requirements.txt
asyncio==3.4.3
beautifulsoup4==4.12.2
blinker==1.7.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
conan==2.0.16
ConfigArgParse==1.7
Faker==22.0.0
fasteners==0.19
Flask==3.0.0
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
gevent==23.9.1
geventhttpclient==2.0.11
google==3.0.0
google-cloud==0.34.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-tools==1.60.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.20.0
MarkupSafe==2.1.3
msgpack==1.0.7
patch-ng==1.17.4
protobuf==4.25.1
psutil==5.9.7
pycparser==2.21
python-dateutil==2.8.2
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
roundrobin==0.0.4
schedule==1.2.1
setuptools==69.0.3
six==1.16.0
soupsieve==2.5
urllib3==1.26.18
Werkzeug==3.0.1
zope.event==5.0
zope.interface==6.1
